+++ CHANGES.md
@@ -5,10 +5,17 @@
 
 - Initial public release.
 
+# Version 0.3.0
+*Date: 2020-09-??*
+
+- Added GetRandomInt() function.
+- Improved error module.
+- Fixed some issues.
+
 # Version 0.2.0
 *Date: 2020-08-19*
 
-- Output methods can optionally omit end-of-line from output
+- Output methods can optionally omit end-of-line from output.
 - Added module for error handling.
 - Added module for audio processing (API-Level 0).